main의 코드를 Controller로 옮기고 일부 수정함
